// src/graphql-types.d.ts
export type Maybe<T> = T | null;
/** All built-in and custom scalars, mapped to their actual values */
export interface Scalars {
	ID: string;
	String: string;
	Boolean: boolean;
	Int: number;
	Float: number;
	// tslint:disable-next-line:jsdoc-format
	/** A date string, such as 2007-12-03, compliant with the ISO 8601 standard for
	 * representation of dates and times using the Gregorian calendar.
	 */
	Date: any;
	/** The `JSON` scalar type represents JSON values as specified by [ECMA-404](http://www.ecma-international.org/publications/files/ECMA-ST/ECMA-404.pdf). */
	JSON: any;
}
// ...
export type Directory = Node & {
	id: Scalars['ID'];
	parent?: Maybe<Node>;
	children: Node[];
	internal: Internal;
	sourceInstanceName?: Maybe<Scalars['String']>;
	absolutePath?: Maybe<Scalars['String']>;
	relativePath?: Maybe<Scalars['String']>;
	extension?: Maybe<Scalars['String']>;
	size?: Maybe<Scalars['Int']>;
	prettySize?: Maybe<Scalars['String']>;
	modifiedTime?: Maybe<Scalars['Date']>;
	accessTime?: Maybe<Scalars['Date']>;
	changeTime?: Maybe<Scalars['Date']>;
	birthTime?: Maybe<Scalars['Date']>;
	root?: Maybe<Scalars['String']>;
	dir?: Maybe<Scalars['String']>;
	base?: Maybe<Scalars['String']>;
	ext?: Maybe<Scalars['String']>;
	name?: Maybe<Scalars['String']>;
	relativeDirectory?: Maybe<Scalars['String']>;
	dev?: Maybe<Scalars['Int']>;
	mode?: Maybe<Scalars['Int']>;
	nlink?: Maybe<Scalars['Int']>;
	uid?: Maybe<Scalars['Int']>;
	gid?: Maybe<Scalars['Int']>;
	rdev?: Maybe<Scalars['Int']>;
	blksize?: Maybe<Scalars['Int']>;
	ino?: Maybe<Scalars['Int']>;
	blocks?: Maybe<Scalars['Int']>;
	atimeMs?: Maybe<Scalars['Float']>;
	mtimeMs?: Maybe<Scalars['Float']>;
	ctimeMs?: Maybe<Scalars['Float']>;
	birthtimeMs?: Maybe<Scalars['Float']>;
	atime?: Maybe<Scalars['Date']>;
	mtime?: Maybe<Scalars['Date']>;
	ctime?: Maybe<Scalars['Date']>;
	birthtime?: Maybe<Scalars['Date']>;
};
// ...
export interface DirectoryConnection {
	totalCount: Scalars['Int'];
	edges: DirectoryEdge[];
	nodes: Directory[];
	pageInfo: PageInfo;
	distinct: Array<Scalars['String']>;
	group: DirectoryGroupConnection[];
}

export interface DirectoryConnectionDistinctArgs {
	field: DirectoryFieldsEnum;
}

export interface DirectoryConnectionGroupArgs {
	skip?: Maybe<Scalars['Int']>;
	limit?: Maybe<Scalars['Int']>;
	field: DirectoryFieldsEnum;
}

export interface DirectoryEdge {
	next?: Maybe<Directory>;
	node: Directory;
	previous?: Maybe<Directory>;
}
// ...
export interface DirectoryGroupConnection {
	totalCount: Scalars['Int'];
	edges: DirectoryEdge[];
	nodes: Directory[];
	pageInfo: PageInfo;
	field: Scalars['String'];
	fieldValue?: Maybe<Scalars['String']>;
}

export interface DirectorySortInput {
	fields?: Maybe<Array<Maybe<DirectoryFieldsEnum>>>;
	order?: Maybe<Array<Maybe<SortOrderEnum>>>;
}

export interface DuotoneGradient {
	highlight?: Maybe<Scalars['String']>;
	shadow?: Maybe<Scalars['String']>;
	opacity?: Maybe<Scalars['Int']>;
}

export enum ExcerptFormats {
	Plain = 'PLAIN',
	Html = 'HTML',
}

export type File = Node & {
	id: Scalars['ID'];
	parent?: Maybe<Node>;
	children: Node[];
	internal: Internal;
	sourceInstanceName?: Maybe<Scalars['String']>;
	absolutePath?: Maybe<Scalars['String']>;
	relativePath?: Maybe<Scalars['String']>;
	extension?: Maybe<Scalars['String']>;
	size?: Maybe<Scalars['Int']>;
	prettySize?: Maybe<Scalars['String']>;
	modifiedTime?: Maybe<Scalars['Date']>;
	accessTime?: Maybe<Scalars['Date']>;
	changeTime?: Maybe<Scalars['Date']>;
	birthTime?: Maybe<Scalars['Date']>;
	root?: Maybe<Scalars['String']>;
	dir?: Maybe<Scalars['String']>;
	base?: Maybe<Scalars['String']>;
	ext?: Maybe<Scalars['String']>;
	name?: Maybe<Scalars['String']>;
	relativeDirectory?: Maybe<Scalars['String']>;
	dev?: Maybe<Scalars['Int']>;
	mode?: Maybe<Scalars['Int']>;
	nlink?: Maybe<Scalars['Int']>;
	uid?: Maybe<Scalars['Int']>;
	gid?: Maybe<Scalars['Int']>;
	rdev?: Maybe<Scalars['Int']>;
	blksize?: Maybe<Scalars['Int']>;
	ino?: Maybe<Scalars['Int']>;
	blocks?: Maybe<Scalars['Int']>;
	atimeMs?: Maybe<Scalars['Float']>;
	mtimeMs?: Maybe<Scalars['Float']>;
	ctimeMs?: Maybe<Scalars['Float']>;
	birthtimeMs?: Maybe<Scalars['Float']>;
	atime?: Maybe<Scalars['Date']>;
	mtime?: Maybe<Scalars['Date']>;
	ctime?: Maybe<Scalars['Date']>;
	birthtime?: Maybe<Scalars['Date']>;
	/** Copy file to static directory and return public url to it */
	publicURL?: Maybe<Scalars['String']>;
	childMarkdownRemark?: Maybe<MarkdownRemark>;
	childImageSharp?: Maybe<ImageSharp>;
};
// ...
export interface FileConnection {
	totalCount: Scalars['Int'];
	edges: FileEdge[];
	nodes: File[];
	pageInfo: PageInfo;
	distinct: Array<Scalars['String']>;
	group: FileGroupConnection[];
}

export interface FileConnectionDistinctArgs {
	field: FileFieldsEnum;
}

export interface FileConnectionGroupArgs {
	skip?: Maybe<Scalars['Int']>;
	limit?: Maybe<Scalars['Int']>;
	field: FileFieldsEnum;
}

export interface FileEdge {
	next?: Maybe<File>;
	node: File;
	previous?: Maybe<File>;
}
// ...
export interface FileGroupConnection {
	totalCount: Scalars['Int'];
	edges: FileEdge[];
	nodes: File[];
	pageInfo: PageInfo;
	field: Scalars['String'];
	fieldValue?: Maybe<Scalars['String']>;
}
// ...
export type ImageSharp = Node & {
	id: Scalars['ID'];
	parent?: Maybe<Node>;
	children: Node[];
	internal: Internal;
	fixed?: Maybe<ImageSharpFixed>;
	resolutions?: Maybe<ImageSharpResolutions>;
	fluid?: Maybe<ImageSharpFluid>;
	sizes?: Maybe<ImageSharpSizes>;
	original?: Maybe<ImageSharpOriginal>;
	resize?: Maybe<ImageSharpResize>;
};
// ...
export interface ImageSharpConnection {
	totalCount: Scalars['Int'];
	edges: ImageSharpEdge[];
	nodes: ImageSharp[];
	pageInfo: PageInfo;
	distinct: Array<Scalars['String']>;
	group: ImageSharpGroupConnection[];
}

export interface ImageSharpConnectionDistinctArgs {
	field: ImageSharpFieldsEnum;
}

export interface ImageSharpConnectionGroupArgs {
	skip?: Maybe<Scalars['Int']>;
	limit?: Maybe<Scalars['Int']>;
	field: ImageSharpFieldsEnum;
}

export interface ImageSharpEdge {
	next?: Maybe<ImageSharp>;
	node: ImageSharp;
	previous?: Maybe<ImageSharp>;
}
// ...
export interface Internal {
	content?: Maybe<Scalars['String']>;
	contentDigest: Scalars['String'];
	description?: Maybe<Scalars['String']>;
	fieldOwners?: Maybe<Array<Maybe<Scalars['String']>>>;
	ignoreType?: Maybe<Scalars['Boolean']>;
	mediaType?: Maybe<Scalars['String']>;
	owner: Scalars['String'];
	type: Scalars['String'];
}
// ...
export interface MarkdownHeading {
	value?: Maybe<Scalars['String']>;
	depth?: Maybe<Scalars['Int']>;
}

export interface MarkdownHeadingFilterInput {
	value?: Maybe<StringQueryOperatorInput>;
	depth?: Maybe<IntQueryOperatorInput>;
}

export interface MarkdownHeadingFilterListInput {
	elemMatch?: Maybe<MarkdownHeadingFilterInput>;
}

export type MarkdownRemark = Node & {
	id: Scalars['ID'];
	parent?: Maybe<Node>;
	children: Node[];
	internal: Internal;
	frontmatter?: Maybe<MarkdownRemarkFrontmatter>;
	excerpt?: Maybe<Scalars['String']>;
	rawMarkdownBody?: Maybe<Scalars['String']>;
	fileAbsolutePath?: Maybe<Scalars['String']>;
	fields?: Maybe<MarkdownRemarkFields>;
	html?: Maybe<Scalars['String']>;
	htmlAst?: Maybe<Scalars['JSON']>;
	excerptAst?: Maybe<Scalars['JSON']>;
	headings?: Maybe<Array<Maybe<MarkdownHeading>>>;
	timeToRead?: Maybe<Scalars['Int']>;
	tableOfContents?: Maybe<Scalars['String']>;
	wordCount?: Maybe<WordCount>;
};
// ...
export interface MarkdownRemarkConnection {
	totalCount: Scalars['Int'];
	edges: MarkdownRemarkEdge[];
	nodes: MarkdownRemark[];
	pageInfo: PageInfo;
	distinct: Array<Scalars['String']>;
	group: MarkdownRemarkGroupConnection[];
}

export interface MarkdownRemarkConnectionDistinctArgs {
	field: MarkdownRemarkFieldsEnum;
}

export interface MarkdownRemarkConnectionGroupArgs {
	skip?: Maybe<Scalars['Int']>;
	limit?: Maybe<Scalars['Int']>;
	field: MarkdownRemarkFieldsEnum;
}

export interface MarkdownRemarkEdge {
	next?: Maybe<MarkdownRemark>;
	node: MarkdownRemark;
	previous?: Maybe<MarkdownRemark>;
}

export interface MarkdownRemarkFields {
	slug?: Maybe<Scalars['String']>;
}
// ...
export interface MarkdownRemarkFrontmatter {
	title?: Maybe<Scalars['String']>;
	date?: Maybe<Scalars['Date']>;
	description?: Maybe<Scalars['String']>;
}
// ...
export interface Node {
	id: Scalars['ID'];
	parent?: Maybe<Node>;
	children: Node[];
	internal: Internal;
}

export interface NodeFilterInput {
	id?: Maybe<StringQueryOperatorInput>;
	parent?: Maybe<NodeFilterInput>;
	children?: Maybe<NodeFilterListInput>;
	internal?: Maybe<InternalFilterInput>;
}

export interface NodeFilterListInput {
	elemMatch?: Maybe<NodeFilterInput>;
}

export interface PageInfo {
	currentPage: Scalars['Int'];
	hasPreviousPage: Scalars['Boolean'];
	hasNextPage: Scalars['Boolean'];
	itemCount: Scalars['Int'];
	pageCount: Scalars['Int'];
	perPage?: Maybe<Scalars['Int']>;
}
// ...
export interface Query {
	file?: Maybe<File>;
	allFile?: Maybe<FileConnection>;
	sitePage?: Maybe<SitePage>;
	allSitePage?: Maybe<SitePageConnection>;
	sitePlugin?: Maybe<SitePlugin>;
	allSitePlugin?: Maybe<SitePluginConnection>;
	site?: Maybe<Site>;
	allSite?: Maybe<SiteConnection>;
	directory?: Maybe<Directory>;
	allDirectory?: Maybe<DirectoryConnection>;
	markdownRemark?: Maybe<MarkdownRemark>;
	allMarkdownRemark?: Maybe<MarkdownRemarkConnection>;
	imageSharp?: Maybe<ImageSharp>;
	allImageSharp?: Maybe<ImageSharpConnection>;
}
// ...
export type Site = Node & {
	id: Scalars['ID'];
	parent?: Maybe<Node>;
	children: Node[];
	internal: Internal;
	siteMetadata?: Maybe<SiteSiteMetadata>;
	port?: Maybe<Scalars['Int']>;
	host?: Maybe<Scalars['String']>;
	pathPrefix?: Maybe<Scalars['String']>;
	polyfill?: Maybe<Scalars['Boolean']>;
	buildTime?: Maybe<Scalars['Date']>;
};

export interface SiteBuildTimeArgs {
	formatString?: Maybe<Scalars['String']>;
	fromNow?: Maybe<Scalars['Boolean']>;
	difference?: Maybe<Scalars['String']>;
	locale?: Maybe<Scalars['String']>;
}

export interface SiteConnection {
	totalCount: Scalars['Int'];
	edges: SiteEdge[];
	nodes: Site[];
	pageInfo: PageInfo;
	distinct: Array<Scalars['String']>;
	group: SiteGroupConnection[];
}

export interface SiteConnectionDistinctArgs {
	field: SiteFieldsEnum;
}

export interface SiteConnectionGroupArgs {
	skip?: Maybe<Scalars['Int']>;
	limit?: Maybe<Scalars['Int']>;
	field: SiteFieldsEnum;
}

export interface SiteEdge {
	next?: Maybe<Site>;
	node: Site;
	previous?: Maybe<Site>;
}
// ...
export interface SiteGroupConnection {
	totalCount: Scalars['Int'];
	edges: SiteEdge[];
	nodes: Site[];
	pageInfo: PageInfo;
	field: Scalars['String'];
	fieldValue?: Maybe<Scalars['String']>;
}

export type SitePage = Node & {
	id: Scalars['ID'];
	parent?: Maybe<Node>;
	children: Node[];
	internal: Internal;
	path?: Maybe<Scalars['String']>;
	jsonName?: Maybe<Scalars['String']>;
	internalComponentName?: Maybe<Scalars['String']>;
	component?: Maybe<Scalars['String']>;
	componentChunkName?: Maybe<Scalars['String']>;
	isCreatedByStatefulCreatePages?: Maybe<Scalars['Boolean']>;
	context?: Maybe<SitePageContext>;
	pluginCreator?: Maybe<SitePlugin>;
	pluginCreatorId?: Maybe<Scalars['String']>;
	componentPath?: Maybe<Scalars['String']>;
};

export interface SitePageConnection {
	totalCount: Scalars['Int'];
	edges: SitePageEdge[];
	nodes: SitePage[];
	pageInfo: PageInfo;
	distinct: Array<Scalars['String']>;
	group: SitePageGroupConnection[];
}

export interface SitePageConnectionDistinctArgs {
	field: SitePageFieldsEnum;
}

export interface SitePageConnectionGroupArgs {
	skip?: Maybe<Scalars['Int']>;
	limit?: Maybe<Scalars['Int']>;
	field: SitePageFieldsEnum;
}

export interface SitePageContext {
	slug?: Maybe<Scalars['String']>;
	previous?: Maybe<SitePageContextPrevious>;
	next?: Maybe<SitePageContextNext>;
}
// ...
export interface SitePageEdge {
	next?: Maybe<SitePage>;
	node: SitePage;
	previous?: Maybe<SitePage>;
}
// ...
export interface SitePageGroupConnection {
	totalCount: Scalars['Int'];
	edges: SitePageEdge[];
	nodes: SitePage[];
	pageInfo: PageInfo;
	field: Scalars['String'];
	fieldValue?: Maybe<Scalars['String']>;
}

export interface SitePageSortInput {
	fields?: Maybe<Array<Maybe<SitePageFieldsEnum>>>;
	order?: Maybe<Array<Maybe<SortOrderEnum>>>;
}

export type SitePlugin = Node & {
	id: Scalars['ID'];
	parent?: Maybe<Node>;
	children: Node[];
	internal: Internal;
	resolve?: Maybe<Scalars['String']>;
	name?: Maybe<Scalars['String']>;
	version?: Maybe<Scalars['String']>;
	pluginOptions?: Maybe<SitePluginPluginOptions>;
	nodeAPIs?: Maybe<Array<Maybe<Scalars['String']>>>;
	browserAPIs?: Maybe<Array<Maybe<Scalars['String']>>>;
	ssrAPIs?: Maybe<Array<Maybe<Scalars['String']>>>;
	pluginFilepath?: Maybe<Scalars['String']>;
	packageJson?: Maybe<SitePluginPackageJson>;
};

export interface SitePluginConnection {
	totalCount: Scalars['Int'];
	edges: SitePluginEdge[];
	nodes: SitePlugin[];
	pageInfo: PageInfo;
	distinct: Array<Scalars['String']>;
	group: SitePluginGroupConnection[];
}

export interface SitePluginConnectionDistinctArgs {
	field: SitePluginFieldsEnum;
}

export interface SitePluginConnectionGroupArgs {
	skip?: Maybe<Scalars['Int']>;
	limit?: Maybe<Scalars['Int']>;
	field: SitePluginFieldsEnum;
}

export interface SitePluginEdge {
	next?: Maybe<SitePlugin>;
	node: SitePlugin;
	previous?: Maybe<SitePlugin>;
}
// ...
export interface SitePluginGroupConnection {
	totalCount: Scalars['Int'];
	edges: SitePluginEdge[];
	nodes: SitePlugin[];
	pageInfo: PageInfo;
	field: Scalars['String'];
	fieldValue?: Maybe<Scalars['String']>;
}

export interface SitePluginPackageJson {
	name?: Maybe<Scalars['String']>;
	description?: Maybe<Scalars['String']>;
	version?: Maybe<Scalars['String']>;
	main?: Maybe<Scalars['String']>;
	author?: Maybe<Scalars['String']>;
	license?: Maybe<Scalars['String']>;
	dependencies?: Maybe<Array<Maybe<SitePluginPackageJsonDependencies>>>;
	devDependencies?: Maybe<Array<Maybe<SitePluginPackageJsonDevDependencies>>>;
	peerDependencies?: Maybe<
		Array<Maybe<SitePluginPackageJsonPeerDependencies>>
	>;
	keywords?: Maybe<Array<Maybe<Scalars['String']>>>;
}
// ...
export interface SitePluginPluginOptions {
	plugins?: Maybe<Array<Maybe<SitePluginPluginOptionsPlugins>>>;
	path?: Maybe<Scalars['String']>;
	name?: Maybe<Scalars['String']>;
	maxWidth?: Maybe<Scalars['Int']>;
	wrapperStyle?: Maybe<Scalars['String']>;
	short_name?: Maybe<Scalars['String']>;
	start_url?: Maybe<Scalars['String']>;
	background_color?: Maybe<Scalars['String']>;
	theme_color?: Maybe<Scalars['String']>;
	display?: Maybe<Scalars['String']>;
	icon?: Maybe<Scalars['String']>;
	pathToConfigModule?: Maybe<Scalars['String']>;
	pathCheck?: Maybe<Scalars['Boolean']>;
}
// ...
export interface SiteSiteMetadata {
	title?: Maybe<Scalars['String']>;
	author?: Maybe<Scalars['String']>;
	description?: Maybe<Scalars['String']>;
	siteUrl?: Maybe<Scalars['String']>;
	social?: Maybe<SiteSiteMetadataSocial>;
}
// ...
export interface WordCount {
	paragraphs?: Maybe<Scalars['Int']>;
	sentences?: Maybe<Scalars['Int']>;
	words?: Maybe<Scalars['Int']>;
}